7. Construction and Design/bad-boids/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "boids"
version = "0.1.0"
description = "A Boids flocking simulation (refactored from bad-boids)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.scripts]
boids = "animate:run_animation"

# ---------------------------------------------------------------------------
# Tool configurations
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true